requirements.txt
asgiref==3.7.2
astroid==3.0.0
attrs==23.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
dill==0.3.7
Django==4.2.6
django-cors-headers==4.2.0
django-filter==23.3
django-model-utils==4.3.1
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.27.2
drf-yasg==1.21.7
idna==3.7
inflection==0.5.1
install==1.3.5
isort==5.12.0
itypes==1.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
Markdown==3.4.4
MarkupSafe==2.1.5
mccabe==0.7.0
openapi-codec==1.3.2
packaging==23.2
Pillow==10.0.1
platformdirs==3.11.0
psycopg2-binary==2.9.9
PyJWT==1.7.1
pylint==3.0.0
python-decouple==3.8
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.3
rpds-py==0.19.0
simplejson==3.19.2
sqlparse==0.4.4
tomli==2.0.1
tomlkit==0.12.1
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.2.2
